refactor retrofit module
