+++ app/build.gradle
@@ -21,6 +21,7 @@ android {
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
+        buildConfigField("String", "BASE_URL", '"https://api-football-v1.p.rapidapi.com/v3/"')
         buildConfigField("String", "API_KEY_VALUE", properties["API_KEY_VALUE"])
     }
 
@@ -50,13 +51,10 @@ android {
 }
 
 dependencies {
-    def room_version = "2.3.0"
-    def nav_version = "2.3.5"
-
     implementation "androidx.core:core-ktx:1.7.0"
-    implementation "androidx.appcompat:appcompat:1.3.1"
+    implementation "androidx.appcompat:appcompat:1.4.0"
     implementation "com.google.android.material:material:1.4.0"
-    implementation "androidx.constraintlayout:constraintlayout:2.1.1"
+    implementation "androidx.constraintlayout:constraintlayout:2.1.2"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.0"
     implementation "androidx.room:room-runtime:$room_version"
@@ -80,9 +78,9 @@ dependencies {
     // Material Design
     implementation "androidx.compose.material:material:1.0.5"
     //Live data
-    implementation "androidx.compose.runtime:runtime-livedata:1.1.0-beta02"
+    implementation "androidx.compose.runtime:runtime-livedata:1.1.0-rc01"
     //Constraints Layout
-    implementation "androidx.constraintlayout:constraintlayout-compose:1.0.0-rc01"
+    implementation "androidx.constraintlayout:constraintlayout-compose:1.0.0-rc02"
     //https://coil-kt.github.io/coil/compose/
     implementation("io.coil-kt:coil-compose:1.4.0")
     //https://androidrepo.com/repo/valentinilk-compose-shimmer
