--- network/src/main/java/com/diegopizzo/network/NetworkConstant.kt
@@ -1,6 +1,0 @@
-package com.diegopizzo.network
-
-object NetworkConstant {
-    const val BASE_URL = "https://api-football-v1.p.rapidapi.com/v3/"
-    const val HEADER_KEY_PARAMETER_NAME = "x-rapidapi-key"
-}
